refactor: rename catch-internal-links.ts -> handle-internal-links.ts
