+++ src/internal-links/handle-internal-links.ts
